+++ crates/models/src/instruction_data.rs
@@ -371,7 +371,6 @@ pub struct WithdrawData {
 /// - `instr_id` - Instr pair id
 /// - `price` - Limit price for a swap
 /// - `amount` - Swaps qty in base crncy
-/// - `ref_fee_rate` - Referral fee for swap orders, 0 by default
 /// - `min_amount_out` - Min amount threshold for trade result, 0 by default
 pub struct SwapData {
     pub tag: u8,
@@ -380,7 +379,6 @@ pub struct SwapData {
     pub instr_id: InstrId,
     pub price: i64,
     pub amount: i64,
-    pub ref_fee_rate: f64,
     pub min_amount_out: i64,
 }
 
@@ -794,4 +792,21 @@ pub struct VmChangeWhitelistData {
     pub padding_u16: u16,
     pub mask: VmMask,
     pub whitelist: [u32; 8],
+}
+
+#[repr(C)]
+#[derive(Clone, Copy, Pod, Zeroable)]
+/// Perp Withdraw Data
+///
+/// **Tag:** `74`
+///
+/// ### Fields
+/// - `instr_id` - Instr pair id
+/// - `amount` - Amount of tokens client wants to move from perp to spot
+pub struct WithdrawSwapFeesData {
+    pub tag: u8,
+    pub padding_u8: u8,
+    pub padding_u16: u16,
+    pub instr_id: InstrId,
+    pub amount: i64,
 }